\name{missForest-package}
\alias{missForest-package}
\docType{package}
\title{
Nonparametric Missing Value Imputation using Random Forest
}
\description{
  'missForest' is used to impute missing values particularly in the case
  of mixed-type data. It can be used to impute continuous and/or
  categorical data including complex interactions and nonlinear
  relations. It yields an out-of-bag (OOB) imputation error estimate.
}
\details{
\tabular{ll}{
Package: \tab missForest\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2012-06-26\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
The main function of the package is \code{\link{missForest}} implementing the
nonparametric missing value imputation. See \code{\link{missForest}} for more
details.
}
\author{
  Daniel J. Stekhoven, <stekhoven@stat.math.ethz.ch>
}
\references{Stekhoven, D.J. and Buehlmann, P. (2012), 'MissForest -
  nonparametric missing value imputation for mixed-type data',
  Bioinformatics, 28(1) 2012, 112-118, doi: 10.1093/bioinformatics/btr597}
\keyword{ package }
\keyword{ nonparametric }
\keyword{ classes }
\keyword{ NA }
